+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 1.0.1 (August 18th, 2021)
+
+Fixes:
+
+- There was no check between successful downloads vs expected downloads. Now, if there is a discrepancy, the script will correctly NOT remove the email.
+- Addition of "forcedelete"-flag to force removal of email even though download may not have been successful.
+
 ## 1.0.0 (August 17th, 2021)
 
 Features: